# RDL.Platform.API/api/openapi.yaml
openapi: 3.1.0
info:
  title: Platform API
  version: 0.7.0
  description: |
    Identidad y tenancy del SaaS de facturación electrónica.
    La organización activa sale SOLO del claim `org_id` del JWT verificado (emitido por el hook de Supabase Auth)
    y de una membresía activa revalidada en cada request. Ningún `organization_id` en body, query, headers o ruta
    cambia el tenant.
    TODO(contracts): alinear con el OpenAPI de Platform del repo de contratos cuando exista.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /v1/me:
    get:
      summary: Perfil del usuario autenticado
      description: La primera llamada da de alta al usuario en core.users (idempotente) y audita `user.provisioned`.
      operationId: getMe
      responses:
        "200":
          description: Perfil
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Me" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /v1/me/memberships:
    get:
      summary: Organizaciones donde el usuario tiene membresía activa
      operationId: listMyMemberships
      responses:
        "200":
          description: Membresías
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MembershipList" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
  /v1/me/active-organization:
    put:
      summary: Selecciona la organización activa
      description: |
        Valida la membresía activa y guarda la selección. El JWT vigente no cambia: el cliente debe refrescar la
        sesión de Supabase para que el nuevo token traiga `org_id`. Responde 404 si no es miembro activo
        (no se revela si la organización existe).
      operationId: selectActiveOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [organizationId]
              properties:
                organizationId: { type: string, format: uuid }
      responses:
        "200":
          description: Seleccionada
          content:
            application/json:
              schema:
                type: object
                required: [activeOrganizationId, tokenRefreshRequired]
                properties:
                  activeOrganizationId: { type: string, format: uuid }
                  tokenRefreshRequired: { type: boolean, const: true }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /v1/organizations:
    post:
      summary: Crea una organización; el creador queda como owner
      description: |
        Idempotente: exige `Idempotency-Key`. Misma clave y mismo cuerpo devuelven 201 con la misma organización
        (header `Idempotent-Replayed: true`); misma clave con otro cuerpo devuelve 422 `idempotency-key-reused`.
        Si el usuario no tenía organización activa, esta queda activa (refresque la sesión para que el token traiga `org_id`).
      operationId: createOrganization
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [legalName, identificationTypeCode, identificationNumber, email]
              properties:
                legalName: { type: string, maxLength: 200 }
                tradeName: { type: string, maxLength: 200 }
                identificationTypeCode:
                  type: string
                  maxLength: 10
                  description: "TODO(contracts): catálogo de tipos de identificación; hoy solo se exige que no esté vacío."
                identificationNumber: { type: string, maxLength: 30 }
                email: { type: string, format: email }
                phone: { type: string, maxLength: 30 }
                timezone: { type: string, description: "Zona IANA. Por defecto America/Costa_Rica." }
      responses:
        "201":
          description: Creada (o repetición idempotente)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Organization" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /v1/organizations/current:
    get:
      summary: Organización activa (todos los roles)
      operationId: getCurrentOrganization
      responses:
        "200":
          description: Organización
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Organization" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Problem" }
    patch:
      summary: Actualiza datos de contacto y presentación (owner, admin)
      description: Campo ausente = sin cambio; "" en tradeName o phone lo borra. Nombre legal e identificación no se editan aquí.
      operationId: updateCurrentOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                tradeName: { type: string, maxLength: 200 }
                email: { type: string, format: email }
                phone: { type: string, maxLength: 30 }
                timezone: { type: string }
      responses:
        "200":
          description: Actualizada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Organization" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /v1/organizations/current/users:
    get:
      summary: Miembros de la organización activa (owner, admin)
      operationId: listMembers
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: cursor, in: query, description: "Valor nextCursor de la página anterior (opaco).", schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [active, suspended] } }
      responses:
        "200":
          description: Página de miembros
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Member" } }
                  nextCursor: { type: [string, "null"] }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /v1/organizations/current/users/{userId}:
    patch:
      summary: Cambia el rol o el estado (baja lógica) de un miembro (owner, admin)
      description: |
        V1: un rol por membresía; asignar un rol reemplaza el anterior.
        Solo un owner puede modificar a otro owner o asignar el rol owner (403 owner-required).
        Nunca puede quedar la organización sin owner activo (409 last-owner).
        Un miembro de otra organización responde 404.
      operationId: updateMember
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              minProperties: 1
              properties:
                role: { $ref: "#/components/schemas/Role" }
                status: { type: string, enum: [active, suspended] }
      responses:
        "200":
          description: Miembro actualizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Member" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT (Supabase Auth, ES256, verificado con JWKS)
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 1 a 255 caracteres ASCII visibles. Se recuerda 24 h.
      schema: { type: string, minLength: 1, maxLength: 255 }
...
  responses:
    Unauthenticated:
      description: Falta el token o no es válido
      headers:
        WWW-Authenticate: { schema: { type: string } }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Problem:
      description: Error (RFC 9457)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          description: |
            Estable por error: urn:rdl:platform:problem:{slug}. Slugs actuales: unauthenticated,
            no-active-organization, membership-inactive, forbidden, not-found, method-not-allowed,
            malformed-request, validation, conflict, email-required, user-disabled, idempotency-key-required,
            idempotency-key-reused, last-owner, owner-required, invitation-expired, invitation-not-pending,
            internal, unavailable.
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        correlationId: { type: string, format: uuid }
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
    Me:
      type: object
      required: [id, email, fullName, status, activeOrganizationId, createdAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        fullName: { type: string }
        status: { type: string, enum: [active, disabled] }
        activeOrganizationId: { type: [string, "null"], format: uuid }
        createdAt: { type: string, format: date-time }
    Organization:
      type: object
      required: [id, legalName, identificationTypeCode, identificationNumber, email, timezone, defaultCurrencyCode, status, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        legalName: { type: string }
        tradeName: { type: string }
        identificationTypeCode: { type: string }
        identificationNumber: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        timezone: { type: string }
        defaultCurrencyCode: { type: string, example: CRC }
        status: { type: string, enum: [active, suspended, closed] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Member:
      type: object
      required: [userId, email, fullName, status, roles, joinedAt]
      properties:
        userId: { type: string, format: uuid }
        email: { type: string, format: email }
        fullName: { type: string }
        status: { type: string, enum: [active, suspended] }
        roles: { type: array, items: { $ref: "#/components/schemas/Role" } }
        joinedAt: { type: string, format: date-time }
...
    Role:
      type: string
      enum: [owner, admin, biller, collector, accountant, read_only]
    MembershipList:
      type: object
      required: [activeOrganizationId, items]
      properties:
        activeOrganizationId: { type: [string, "null"], format: uuid }
        items:
          type: array
          items:
            type: object
            required: [organizationId, legalName, organizationStatus, status, roles, joinedAt, isActive]
            properties:
              organizationId: { type: string, format: uuid }
              legalName: { type: string }
              tradeName: { type: string }
              organizationStatus: { type: string, enum: [active, suspended, closed] }
              status: { type: string, enum: [active] }
              roles:
                type: array
                description: V1 asigna un solo rol por membresía; la lista deja lista la API para varios.
                items: { $ref: "#/components/schemas/Role" }
              joinedAt: { type: string, format: date-time }
              isActive: { type: boolean, description: Es la organización activa del usuario }